refactor(project): 서비스 기능별 분리
