+++ src/main/java/com/moa/backend/domain/project/service/ProjectCommandService.java
@@ -1,0 +1,9 @@
+package com.moa.backend.domain.project.service;
+
+import com.moa.backend.domain.project.dto.CreateProjectRequest;
+import com.moa.backend.domain.project.dto.CreateProjectResponse;
+
+public interface ProjectCommandService {
+
+    CreateProjectResponse createProject(Long userId, CreateProjectRequest request);
+}
